backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
matplotlib==3.8.2
scipy==1.11.4
scikit-learn==1.3.2
plotly==5.18.0
beautifulsoup4==4.12.2
lxml==4.9.4
openpyxl==3.1.5
